storage: fix the problem with reading of EERPOM in large chunks
The reading fails if more than 4096 bytes of data is read at once. The
problem seems to be in the limitation/feature of sysfs interface, which
returns data to userpsace in chunks of PAGE_SIZE:
https://www.kernel.org/doc/Documentation/filesystems/sysfs.txt

This patch reworks the read routing to receive data from kernel in loop until
all data is read.